rsa::algorithms::rsa

Function rsa_decrypt_and_check

pub fn rsa_decrypt_and_check<R: TryCryptoRng + ?Sized>(
    priv_key: &impl PrivateKeyParts,
    rng: Option<&mut R>,
    c: &BoxedUint,
) -> Result<BoxedUint>

⚠️ Performs raw RSA decryption with no padding.

Returns a plaintext BoxedUint. Performs RSA blinding if an Rng is passed. This will also check for errors in the CRT computation.

c must have the same bits_precision as the RSA key modulus.

☢️️ WARNING: HAZARDOUS API ☢️

Use this function with great care! Raw RSA should never be used without an appropriate padding or signature scheme. See the [module-level documentation][crate::hazmat] for more information.