rsa::key

Struct RsaPrivateKey

pub struct RsaPrivateKey {
    pubkey_components: RsaPublicKey,
    pub(crate) d: BoxedUint,
    pub(crate) primes: Vec<BoxedUint>,
    pub(crate) precomputed: Option<PrecomputedValues>,
}

Represents a whole RSA key, public and private parts.

Fields
pubkey_components: RsaPublicKey

Public components of the private key.

d: BoxedUint

Private exponent

primes: Vec<BoxedUint>

Prime factors of N, contains >= 2 elements.

precomputed: Option<PrecomputedValues>

Precomputed values to speed up private operations

Implementations

impl RsaPrivateKey

const EXP: u64 = 65537

Default exponent for RSA keys.

const MIN_SIZE: u32 = 1024

Minimum size of the modulus n in bits. Currently only applies to keygen.

pub fn new<R: CryptoRng + ?Sized>(rng: &mut R, bit_size: usize) -> Result<Self>

Generate a new RSA key pair with a modulus of the given bit size using the passed in rng.

Errors

If bit_size is lower than the minimum 1024-bits.

pub fn new_with_exp<R: CryptoRng + ?Sized>( rng: &mut R, bit_size: usize, exp: BoxedUint, ) -> Result<RsaPrivateKey>

Generate a new RSA key pair of the given bit size and the public exponent using the passed in rng.

Unless you have specific needs, you should use RsaPrivateKey::new instead.

fn from_components_inner( n: BoxedUint, e: BoxedUint, d: BoxedUint, primes: Vec<BoxedUint>, ) -> Result<RsaPrivateKey>

Private helper function that constructs an RSA key pair from components WITHOUT performing any validation or precomputation.

This is the shared implementation used by from_components and from_components_with_large_exponent.

Callers are responsible for:

Validating the key (to ensure precomputation won’t fail)

Calling precompute() after validation

pub fn from_components( n: BoxedUint, e: BoxedUint, d: BoxedUint, primes: Vec<BoxedUint>, ) -> Result<RsaPrivateKey>

Constructs an RSA key pair from individual components:

n: RSA modulus

e: public exponent (i.e. encrypting exponent)

d: private exponent (i.e. decrypting exponent)

primes: prime factors of n: typically two primes p and q. More than two primes can be provided for multiprime RSA, however this is generally not recommended. If no primes are provided, a prime factor recovery algorithm will be employed to attempt to recover the factors (as described in NIST SP 800-56B Revision 2 Appendix C.2). This algorithm only works if there are just two prime factors p and q (as opposed to multiprime), and e is between 2^16 and 2^256.

pub fn from_p_q( p: BoxedUint, q: BoxedUint, public_exponent: BoxedUint, ) -> Result<RsaPrivateKey>

Constructs an RSA key pair from its two primes p and q.

This will rebuild the private exponent and the modulus.

Private exponent will be rebuilt using the method defined in NIST 800-56B Section 6.2.1.

pub fn from_primes( primes: Vec<BoxedUint>, public_exponent: BoxedUint, ) -> Result<RsaPrivateKey>

Constructs an RSA key pair from its primes.

This will rebuild the private exponent and the modulus.

pub fn as_public_key(&self) -> &RsaPublicKey

Get the public key from the private key.

Specific alternative to AsRef::as_ref.

pub fn to_public_key(&self) -> RsaPublicKey

Get the public key from the private key, cloning n and e.

Generally this is not needed since RsaPrivateKey implements the PublicKey trait, but it can occasionally be useful to discard the private information entirely.

pub fn precompute(&mut self) -> Result<()>

Performs some calculations to speed up private key operations.

pub fn clear_precomputed(&mut self)

Clears precomputed values by setting to None

pub fn crt_coefficient(&self) -> Option<BoxedUint>

Compute CRT coefficient: (1/q) mod p.

pub fn validate(&self) -> Result<()>

Performs basic sanity checks on the key. Returns Ok(()) if everything is good, otherwise an appropriate error.

pub fn decrypt<P: PaddingScheme>( &self, padding: P, ciphertext: &[u8], ) -> Result<Vec<u8>>

Decrypt the given message.

pub fn decrypt_blinded<R: CryptoRng + ?Sized, P: PaddingScheme>( &self, rng: &mut R, padding: P, ciphertext: &[u8], ) -> Result<Vec<u8>>

Decrypt the given message.

Uses rng to blind the decryption process.

pub fn sign<S: SignatureScheme>( &self, padding: S, digest_in: &[u8], ) -> Result<Vec<u8>>

Sign the given digest.

pub fn sign_with_rng<R: CryptoRng + ?Sized, S: SignatureScheme>( &self, rng: &mut R, padding: S, digest_in: &[u8], ) -> Result<Vec<u8>>

Sign the given digest using the provided rng, which is used in the following ways depending on the SignatureScheme:

Pkcs1v15Sign padding: uses the RNG to mask the private key operation with random blinding, which helps mitigate sidechannel attacks.

Pss always requires randomness. Use Pss::new for a standard RSASSA-PSS signature, or Pss::new_blinded for RSA-BSSA blind signatures.