rsa::pkcs1v15

Struct SigningKey

pub struct SigningKey<D>
where
    D: Digest,
{
    inner: RsaPrivateKey,
    prefix: Vec<u8>,
    phantom: PhantomData<D>,
}

Signing key for RSASSA-PKCS1-v1_5 signatures as described in RFC8017  8.2.

Fields
inner: RsaPrivateKeyprefix: Vec<u8>phantom: PhantomData<D>
Implementations

impl<D> SigningKey<D>
where D: Digest + AssociatedOid,

pub fn new(key: RsaPrivateKey) -> Self

Create a new signing key with a prefix for the digest D.

pub fn random<R: CryptoRng + ?Sized>( rng: &mut R, bit_size: usize, ) -> Result<Self>

Generate a new signing key with a prefix for the digest D.

impl<D> SigningKey<D>
where D: Digest,

pub fn new_unprefixed(key: RsaPrivateKey) -> Self

Create a new signing key from the give RSA private key with an empty prefix.

Note: unprefixed signatures are uncommon

In most cases you’ll want to use SigningKey::new.

pub fn random_unprefixed<R: CryptoRng + ?Sized>( rng: &mut R, bit_size: usize, ) -> Result<Self>

Generate a new signing key with an empty prefix.