rsa::pss

Function sign_pss_with_salt

fn sign_pss_with_salt<T: CryptoRngCore>(
    blind_rng: Option<&mut T>,
    priv_key: &RsaPrivateKey,
    hashed: &[u8],
    salt: &[u8],
    digest: &mut dyn DynDigest,
) -> Result<Vec<u8>>

signPSSWithSalt calculates the signature of hashed using PSS with specified salt.

Note that hashed must be the result of hashing the input message using the given hash function. salt is a random sequence of bytes whose length will be later used to verify the signature.