Struct rsa::pss::SigningKey

pub struct SigningKey<D>
where
    D: Digest,
{ /* private fields */ }

Signing key for producing RSASSA-PSS signatures as described in RFC8017  8.1.

Implementations

impl<D> SigningKey<D>
where D: Digest,

pub fn new(key: RsaPrivateKey) -> Self

Create a new RSASSA-PSS signing key. Digest output size is used as a salt length.

pub fn new_with_salt_len(key: RsaPrivateKey, salt_len: usize) -> Self

Create a new RSASSA-PSS signing key with a salt of the given length.

pub fn random<R: CryptoRngCore + ?Sized>( rng: &mut R, bit_size: usize ) -> Result<Self>

Generate a new random RSASSA-PSS signing key. Digest output size is used as a salt length.

pub fn random_with_salt_len<R: CryptoRngCore + ?Sized>( rng: &mut R, bit_size: usize, salt_len: usize ) -> Result<Self>

Generate a new random RSASSA-PSS signing key with a salt of the given length.

pub fn salt_len(&self) -> usize

Return specified salt length for this key