Struct rsa::RsaPublicKey

pub struct RsaPublicKey { /* private fields */ }

Represents the public part of an RSA key.

Implementations

impl RsaPublicKey

pub const MIN_PUB_EXPONENT: u64 = 2u64

Minimum value of the public exponent e.

pub const MAX_PUB_EXPONENT: u64 = 8_589_934_591u64

Maximum value of the public exponent e.

pub const MAX_SIZE: usize = 4_096usize

Maximum size of the modulus n in bits.

pub fn new(n: BigUint, e: BigUint) -> Result<Self>

Create a new public key from its components.

This function accepts public keys with a modulus size up to 4096-bits, i.e. RsaPublicKey::MAX_SIZE.

pub fn new_with_max_size( n: BigUint, e: BigUint, max_size: usize ) -> Result<Self>

Create a new public key from its components.

pub fn new_unchecked(n: BigUint, e: BigUint) -> Self

Create a new public key, bypassing checks around the modulus and public exponent size.

This method is not recommended, and only intended for unusual use cases. Most applications should use RsaPublicKey::new or RsaPublicKey::new_with_max_size instead.