rsa::algorithms

Module pad

Special handling for converting the BigUint to u8 vectors

Functions

left_pad 🔒

Returns a new vector of the given length, with 0s left padded.

uint_to_be_pad 🔒

Converts input to the new vector of the given length, using BE and with 0s left padded.

uint_to_zeroizing_be_pad 🔒

Converts input to the new vector of the given length, using BE and with 0s left padded.