rsa::algorithms::rsa

Function recover_primes

pub fn recover_primes(
    n: &BigUint,
    e: &BigUint,
    d: &BigUint,
) -> Result<(BigUint, BigUint)>

The following (deterministic) algorithm also recovers the prime factors p and q of a modulus n, given the public exponent e and private exponent d using the method described in NIST 800-56B Appendix C.2.