rsa

Module pkcs1v15

PKCS#1 v1.5 support as described in RFC8017  8.2.

Usage

See code example in the toplevel rustdoc.

Modules

decrypting_key 🔒

encrypting_key 🔒

oid 🔒

signature 🔒

signing_key 🔒

verifying_key 🔒

Structs

DecryptingKey

Decryption key for PKCS#1 v1.5 decryption as described in RFC8017  7.2.

EncryptingKey

Encryption key for PKCS#1 v1.5 encryption as described in RFC8017  7.2.

Pkcs1v15Encrypt

Encryption using PKCS#1 v1.5 padding.

Pkcs1v15Sign

RSASSA-PKCS1-v1_5: digital signatures using PKCS#1 v1.5 padding.

Signature

RSASSA-PKCS1-v1_5 signatures as described in RFC8017  8.2.

SigningKey

Signing key for RSASSA-PKCS1-v1_5 signatures as described in RFC8017  8.2.

VerifyingKey

Verifying key for RSASSA-PKCS1-v1_5 signatures as described in RFC8017  8.2.

Traits

RsaSignatureAssociatedOid

A trait which associates an RSA-specific OID with a type.

Functions

decrypt 🔒

Decrypts a plaintext using RSA and the padding scheme from PKCS#1 v1.5.

encrypt 🔒

Encrypts the given message with RSA and the padding scheme from PKCS#1 v1.5. The message must be no longer than the length of the public modulus minus 11 bytes.

sign 🔒

Calculates the signature of hashed using RSASSA-PKCS1-V1_5-SIGN from RSA PKCS#1 v1.5. Note that hashed must be the result of hashing the input message using the given hash function. If hash is None, hashed is signed directly. This isn’t advisable except for interoperability.

verify 🔒

Verifies an RSA PKCS#1 v1.5 signature.