Struct rsa::pkcs1v15::SigningKey

pub struct SigningKey<D>where
    D: Digest,{ /* private fields */ }

Signing key for PKCS#1 v1.5 signatures as described in RFC8017  8.2.

Implementations

impl<D> SigningKey<D>where D: Digest,

pub fn new(key: RsaPrivateKey) -> Self

Create a new signing key from the give RSA private key with an empty prefix.

Note: unprefixed signatures are uncommon

In most cases you’ll want to use SigningKey::new_with_prefix.

pub fn random<R: CryptoRngCore + ?Sized>( rng: &mut R, bit_size: usize ) -> Result<Self>

Generate a new signing key.

impl<D> SigningKey<D>where D: Digest + AssociatedOid,

pub fn new_with_prefix(key: RsaPrivateKey) -> Self

Create a new signing key with a prefix for the digest D.

pub fn random_with_prefix<R: CryptoRngCore + ?Sized>( rng: &mut R, bit_size: usize ) -> Result<Self>

Generate a new signing key with a prefix for the digest D.