Struct rsa::pss::Pss

pub struct Pss {
    pub blinded: bool,
    pub digest: Box<dyn DynDigest + Send + Sync>,
    pub salt_len: Option<usize>,
}

Digital signatures using PSS padding.

Fields
blinded: bool

Create blinded signatures.

digest: Box<dyn DynDigest + Send + Sync>

Digest type to use.

salt_len: Option<usize>

Salt length.

Implementations

impl Pss

pub fn new<T: 'static + Digest + DynDigest + Send + Sync>() -> Self

New PSS padding for the given digest.

pub fn new_with_salt<T: 'static + Digest + DynDigest + Send + Sync>( len: usize ) -> Self

New PSS padding for the given digest with a salt value of the given length.

pub fn new_blinded<T: 'static + Digest + DynDigest + Send + Sync>() -> Self

New PSS padding for blinded signatures (RSA-BSSA) for the given digest.

pub fn new_blinded_with_salt<T: 'static + Digest + DynDigest + Send + Sync>( len: usize ) -> Self

New PSS padding for blinded signatures (RSA-BSSA) for the given digest with a salt value of the given length.