Struct rsa::RsaPrivateKey

pub struct RsaPrivateKey { /* private fields */ }

Represents a whole RSA key, public and private parts.

Implementations

impl RsaPrivateKey

pub fn new<R: CryptoRngCore + ?Sized>( rng: &mut R, bit_size: usize ) -> Result<RsaPrivateKey>

Generate a new Rsa key pair of the given bit size using the passed in rng.

pub fn new_with_exp<R: CryptoRngCore + ?Sized>( rng: &mut R, bit_size: usize, exp: &BigUint ) -> Result<RsaPrivateKey>

Generate a new RSA key pair of the given bit size and the public exponent using the passed in rng.

Unless you have specific needs, you should use RsaPrivateKey::new instead.

pub fn from_components( n: BigUint, e: BigUint, d: BigUint, primes: Vec<BigUint> ) -> Result<RsaPrivateKey>

Constructs an RSA key pair from the individual components.

pub fn to_public_key(&self) -> RsaPublicKey

Get the public key from the private key, cloning n and e.

Generally this is not needed since RsaPrivateKey implements the PublicKey trait, but it can occasionally be useful to discard the private information entirely.

pub fn precompute(&mut self) -> Result<()>

Performs some calculations to speed up private key operations.

pub fn clear_precomputed(&mut self)

Clears precomputed values by setting to None

pub fn dp(&self) -> Option<&BigUint>

Returns the precomputed dp value, D mod (P-1)

pub fn dq(&self) -> Option<&BigUint>

Returns the precomputed dq value, D mod (Q-1)

pub fn qinv(&self) -> Option<&BigInt>

Returns the precomputed qinv value, Q^-1 mod P

pub fn d(&self) -> &BigUint

Returns the private exponent of the key.

pub fn primes(&self) -> &[BigUint]

Returns the prime factors.

pub fn crt_coefficient(&self) -> Option<BigUint>

Compute CRT coefficient: (1/q) mod p.

pub fn validate(&self) -> Result<()>

Performs basic sanity checks on the key. Returns Ok(()) if everything is good, otherwise an appropriate error.

pub fn decrypt<P: PaddingScheme>( &self, padding: P, ciphertext: &[u8] ) -> Result<Vec<u8>>

Decrypt the given message.

pub fn decrypt_blinded<R: CryptoRngCore, P: PaddingScheme>( &self, rng: &mut R, padding: P, ciphertext: &[u8] ) -> Result<Vec<u8>>

Decrypt the given message.

Uses rng to blind the decryption process.

pub fn sign<S: SignatureScheme>( &self, padding: S, digest_in: &[u8] ) -> Result<Vec<u8>>

Sign the given digest.

pub fn sign_with_rng<R: CryptoRngCore, S: SignatureScheme>( &self, rng: &mut R, padding: S, digest_in: &[u8] ) -> Result<Vec<u8>>

Sign the given digest using the provided rng, which is used in the following ways depending on the SignatureScheme:

Pkcs1v15Sign padding: uses the RNG to mask the private key operation with random blinding, which helps mitigate sidechannel attacks.

Pss always requires randomness. Use Pss::new for a standard RSASSA-PSS signature, or Pss::new_blinded for RSA-BSSA blind signatures.