Struct rsa::RsaPrivateKey

pub struct RsaPrivateKey { /* private fields */ }

Represents a whole RSA key, public and private parts.

Implementations

impl RsaPrivateKey

pub fn new<R: Rng>(rng: &mut R, bit_size: usize) -> Result<RsaPrivateKey>

Generate a new Rsa key pair of the given bit size using the passed in rng.

pub fn new_with_exp<R: Rng>(
rng: &mut R,
bit_size: usize,
exp: &BigUint
) -> Result<RsaPrivateKey>

Generate a new RSA key pair of the given bit size and the public exponent using the passed in rng.

Unless you have specific needs, you should use RsaPrivateKey::new instead.

pub fn from_components(
n: BigUint,
e: BigUint,
d: BigUint,
primes: Vec<BigUint>
) -> RsaPrivateKey

Constructs an RSA key pair from the individual components.

pub fn to_public_key(&self) -> RsaPublicKey

Get the public key from the private key, cloning n and e.

Generally this is not needed since RsaPrivateKey implements the PublicKey trait, but it can occationally be useful to discard the private information entirely.

pub fn precompute(&mut self) -> Result<()>

Performs some calculations to speed up private key operations.

pub fn d(&self) -> &BigUint

Returns the private exponent of the key.

pub fn primes(&self) -> &[BigUint]

Returns the prime factors.

pub fn validate(&self) -> Result<()>

Performs basic sanity checks on the key. Returns Ok(()) if everything is good, otherwise an approriate error.

pub fn decrypt(
&self,
padding: PaddingScheme,
ciphertext: &[u8]
) -> Result<Vec<u8>>

Decrypt the given message.

pub fn decrypt_blinded<R: Rng>(
&self,
rng: &mut R,
padding: PaddingScheme,
ciphertext: &[u8]
) -> Result<Vec<u8>>

Decrypt the given message.

Uses rng to blind the decryption process.

pub fn sign(&self, padding: PaddingScheme, digest_in: &[u8]) -> Result<Vec<u8>>

Sign the given digest.

pub fn sign_blinded<R: Rng>(
&self,
rng: &mut R,
padding: PaddingScheme,
digest_in: &[u8]
) -> Result<Vec<u8>>

Sign the given digest.

Use rng for blinding.