rsa::traits

Trait EncryptingKeypair

pub trait EncryptingKeypair {
    type EncryptingKey: Clone;

    // Required method
    fn encrypting_key(&self) -> Self::EncryptingKey;
}

Encryption keypair with an associated encryption key.

Required Associated Types

type EncryptingKey: Clone

Encrypting key type for this keypair.

Required Methods

fn encrypting_key(&self) -> Self::EncryptingKey

Get the encrypting key which can encrypt messages to be decrypted by the decryption key portion of this keypair.